tried fresh start, separated int negative treatment from regular treatment

+++ srcs/libft/includes/libft.h
@@ -6,7 +6,7 @@
 /*   By: rponsonn <rponsonn@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2021/01/11 16:54:10 by rponsonn          #+#    #+#             */
-/*   Updated: 2021/04/18 21:56:00 by rponsonn         ###   ########.fr       */
+/*   Updated: 2021/04/20 13:27:23 by rponsonn         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -52,7 +52,7 @@ char			*ft_substr(const char *s, unsigned int start, size_t len);
 char			*ft_strjoin(const char *s1, char const *s2);
 char			*ft_strtrim(const char *s1, const char *s2);
 char			**ft_split(const char *s, char c);
-char			*ft_itoa(int n);
+char			*ft_itoa(long int n);
 char			*ft_strmapi(const char *s, char (*f)(unsigned int, char));
 void			ft_putchar_fd(char c, int fd);
 void			ft_putstr_fd(char *s, int fd);
